% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add2vars.R
\name{add2vars}
\alias{add2vars}
\title{add2vars}
\usage{
add2vars(x, y)
}
\arguments{
\item{x}{first numerical input}

\item{y}{second numerical input}
}
\value{
sum of x & y
}
\description{
add2vars
}
\examples{
add2vars(2, 3)

}
