% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_risk.R
\name{predict_risk.data.frame}
\alias{predict_risk.data.frame}
\title{Predict risk using a data frame}
\usage{
\method{predict_risk}{data.frame}(x, ...)
}
\arguments{
\item{x}{A data frame with columns: sex, wbc, treatment.}

\item{...}{Not used.}
}
\value{
A numeric vector of risk predictions.
}
\description{
Predict risk using a data frame
}
