% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_risk.R
\name{predict_risk.default}
\alias{predict_risk.default}
\title{Predict risk using individual vectors}
\usage{
\method{predict_risk}{default}(x = NULL, sex, wbc, treatment, ...)
}
\arguments{
\item{x}{Placeholder, ignored.}

\item{sex}{Character vector of "male" or "female".}

\item{wbc}{Numeric vector of white blood cell counts.}

\item{treatment}{Character vector of "standard" or "new".}

\item{...}{Unused.}
}
\value{
A numeric vector of risk predictions.
}
\description{
Predict risk using individual vectors
}
