% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_risk.R
\name{predict_risk}
\alias{predict_risk}
\title{Predict risk of Leukemia remission}
\usage{
predict_risk(x, ...)
}
\arguments{
\item{x}{Either a data.frame or values for sex, wbc, treatment.}

\item{...}{Additional arguments (e.g., sex, wbc, treatment if x is not a data frame).}
}
\value{
A numeric vector of relative risk predictions.
}
\description{
Dispatches to appropriate method based on input type.
}
