% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-programTable.R
\docType{data}
\name{programTable}
\alias{programTable}
\title{RCC Program Table}
\format{
A tibble with 12 rows and 4 variables:
\describe{
\item{type}{Type/category of the program (e.g., Paramedical Courses, Outreach Programs)}
\item{name}{Name of the program}
\item{duration}{Duration of the program (e.g., "1 year", "not specified")}
\item{description}{Brief description of the program}
}
}
\source{
\url{https://www.rcctvm.gov.in}
}
\usage{
programTable
}
\description{
A dataset describing various educational and outreach programs conducted by the Regional Cancer Centre (RCC) Thiruvananthapuram.
}
\keyword{datasets}
